Switching between auth modes
Toggle between 'Sign up' and 'Login' button depending on whether the application is in login mode.

Toggle between 'Switch to Sign Up' and 'Switch to Login' button depending on whether the application is in login mode.

Toggle between login and sign up mode when the button is pressed to switch modes.