Handle e-mail and password inputs for login and sign up form
Connect e-mail input to ngModel. This input is required and in the format of an e-mail.

Connect password input to ngModel. This input is required and in the format of a password. The password needs to have a minimum length of at least 6 characters.

Disable the login/sign up button if the e-mail or password input is invalid.

Reset the login/sign up form once it is submitted.